# android/WellOffice/app/src/main/java/it/orbyta/welloffice/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WellOffice",
    "version": "1.0"
  },
  "paths": {
    "/api/Parameter": {
      "get": {
        "tags": [
          "Parameter"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Parameter"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Parameter"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Parameter"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Parameter"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Parameter"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Parameter"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Parameter"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              }
            }
          }
        }
      }
    },
    "/api/Parameter/{id}": {
      "get": {
        "tags": [
          "Parameter"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Parameter"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/RemediationAction": {
      "get": {
        "tags": [
          "RemediationAction"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RemediationAction"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationAction"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationAction"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RemediationAction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationAction"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationAction"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationAction"
                }
              }
            }
          }
        }
      }
    },
    "/api/RemediationAction/{id}": {
      "get": {
        "tags": [
          "RemediationAction"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationAction"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationAction"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RemediationAction"
                }
              }
            }
          }
        }
      }
    },
    "/api/RemediationAction/sensor/{sensorId}": {
      "get": {
        "tags": [
          "RemediationAction"
        ],
        "parameters": [
          {
            "name": "sensorId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/RemediationAction/sensor/{sensorId}/latest": {
      "get": {
        "tags": [
          "RemediationAction"
        ],
        "parameters": [
          {
            "name": "sensorId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "count",
            "in": "query",
            "style": "form",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RemediationAction"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Sensor/{id}": {
      "get": {
        "tags": [
          "Sensor"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Sensor"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Sensor/parameter/{parameterId}": {
      "get": {
        "tags": [
          "Sensor"
        ],
        "parameters": [
          {
            "name": "parameterId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sensor"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sensor"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Sensor"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Threshold/parameter/{parameterId}": {
      "get": {
        "tags": [
          "Threshold"
        ],
        "parameters": [
          {
            "name": "parameterId",
            "in": "path",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Threshold"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Threshold"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Threshold"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Parameter": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "unitMeasure": {
            "$ref": "#/components/schemas/UnitMeasure"
          },
          "sensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sensor"
            },
            "nullable": true
          },
          "thresholds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Threshold"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "RemediationAction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sensorId": {
            "type": "string",
            "format": "uuid"
          },
          "action": {
            "type": "string",
            "nullable": true
          },
          "actionDate": {
            "type": "string",
            "format": "date-time"
          },
          "sensor": {
            "$ref": "#/components/schemas/Sensor"
          }
        },
        "additionalProperties": false
      },
      "Room": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "footage": {
            "type": "number",
            "format": "double"
          },
          "ceilingHeight": {
            "type": "number",
            "format": "double"
          },
          "sensors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sensor"
            },
            "nullable": true
          },
          "thresholds": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Threshold"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Sensor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "parameterId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "isActive": {
            "type": "boolean"
          },
          "parameter": {
            "$ref": "#/components/schemas/Parameter"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          },
          "sensorData": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SensorData"
            },
            "nullable": true
          },
          "remediationActions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RemediationAction"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "SensorData": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sensorId": {
            "type": "string",
            "format": "uuid"
          },
          "value": {
            "type": "number",
            "format": "double"
          },
          "detectionDate": {
            "type": "string",
            "format": "date-time"
          },
          "sensor": {
            "$ref": "#/components/schemas/Sensor"
          }
        },
        "additionalProperties": false
      },
      "Threshold": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "parameterId": {
            "type": "string",
            "format": "uuid"
          },
          "roomId": {
            "type": "string",
            "format": "uuid"
          },
          "optimalMinValue": {
            "type": "number",
            "format": "double"
          },
          "optimalMaxValue": {
            "type": "number",
            "format": "double"
          },
          "acceptableMinValue": {
            "type": "number",
            "format": "double"
          },
          "acceptableMaxValue": {
            "type": "number",
            "format": "double"
          },
          "parameter": {
            "$ref": "#/components/schemas/Parameter"
          },
          "room": {
            "$ref": "#/components/schemas/Room"
          }
        },
        "additionalProperties": false
      },
      "UnitMeasure": {
        "enum": [
          0,
          1,
          2,
          3,
          4,
          5,
          6,
          7,
          8
        ],
        "type": "integer",
        "format": "int32"
      },
...
    }
  }
}