# aspnet-core/WellOffice/Models/UnitMeasure.cs
namespace WellOffice.Models;

public enum UnitMeasure
{
    Temperature,    // Celsius
    Humidity,       // Percentage
    CO2,           // ppm (parts per million)
    VOC,           // ppb (parts per billion)
    Pressure,      // hPa
    Light,         // lux
    Noise,         // dB
    AirFlow,       // m³/h
    Occupancy,     // persone/m²
    AirQuality,    // IAQ (Indoor Air Quality)
    Other
} 